Moved camera-specific functions into their own file
